User commit

+++ src/main/java/com/example/demo/entity/student.java
@@ -12,13 +12,13 @@ import jakarta.persistence.Id;
 public class Student {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String name;
     private String email;
-    public int getId() {
+    public Long getId() {
         return id;
     }
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
     public String getName() {
@@ -33,7 +33,7 @@ public class Student {
     public void setEmail(String email) {
         this.email = email;
     }
-    public Student(int id, String name, String email, LocalDate dob, float cgpa) {
+    public Student(Long id, String name, String email, LocalDate dob, float cgpa) {
         this.id = id;
         this.name = name;
         this.email = email;
